Changes according to code review. EmptyResult created. Type of EngineSize changed to double.

+++ SpeedyWheelsSales/src/SpeedyWheelsSales.Domain/Entities/Car.cs
@@ -8,7 +8,7 @@ public class Car
     public string Model { get; set; }
     public decimal Price { get; set; }
     public int Mileage { get; set; }
-    public decimal EngineSize { get; set; }
+    public double EngineSize { get; set; }
     public double FuelConsumption { get; set; }
     public string Vin { get; set; }
     public string Plates { get; set; }
